backend: rename webApp.js to web.js, extends docs
